+++ packages/react-router/src/link.tsx
@@ -257,7 +257,7 @@ export function useLinkProps<
   } = options as typeof options & { to?: string }
 
   // eslint-disable-next-line react-hooks/rules-of-hooks
-  const isHydrated = useHydrated()
+  const isHydrated = useHydrated(!!activeOptions?.includeHash)
 
   // eslint-disable-next-line react-hooks/rules-of-hooks
   const [stableSearch, stableParams, stableActiveOptions] = useStableValues(